fix: decided to separate tests for consistency, update readme to reflect
