+++ starters/vue3-xstate-css/cypress.config.ts
@@ -4,7 +4,7 @@ export default defineConfig({
   viewportHeight: 720,
   viewportWidth: 1280,
   component: {
-    specPattern: 'src/**/__tests__/*.{cy,spec}.{js,ts,jsx,tsx}',
+    specPattern: 'src/tests/*.{cy,spec}.{js,ts,jsx,tsx}',
     devServer: {
       framework: 'vue',
       bundler: 'vite',
